Refactor src/app/pages/Dashboard/Content/DragLayer/SelectedSandboxItems/AnimatedSandboxItem to Typescript and functional component with hook
